Mejoras en los datos de salida.

+++ EjercicioExtExcTest/EjercicioExtExcTest/SeveralOperations.cs
@@ -13,7 +13,7 @@ namespace EjercicioExtExcTest
             return number / 0;
         }
 
-        public static float DivideTwoNumbers(int firstNumber, int secondNumber)
+        public static decimal DivideTwoNumbers(decimal firstNumber, decimal secondNumber)
         {
             return firstNumber / secondNumber;
         }
